+++ game-frontend/projects/owge-widgets/src/public-api.ts
@@ -18,3 +18,4 @@ export * from './lib/owge-widgets.module';
 export * from './lib/pipes/ui-icon.pipe';
 export * from './lib/services/display.service';
 export * from './lib/types/widget-filter.type';
+export * from './lib/utils/widget-filter.util';